feat: 공개캡슐 디테일 usecase

+++ frontend/ARchive/app/src/main/java/com/droidblossom/archive/data/source/remote/api/PublicService.kt
@@ -24,7 +24,7 @@ interface PublicService {
 
     @GET("public/capsules/{capsule_id}/detail")
     suspend fun getPublicCapsuleDetailApi(
-        @Path("capsule_id") capsuleId : Int,
+        @Path("capsule_id") capsuleId : Long,
     ) : Response<ResponseBody<CapsuleDetailResponseDto>>
 
 }
